Predict anticancer peptides

Recognizes anticancer peptides using the CancerGram algorithm.

# S3 method for cancergram_model
predict(object, newdata, ...)

Arguments

object
cancergram_model object.

newdata
list of sequences (for example as given by read_fasta or read_txt).

...
further arguments passed to or from other methods.

Value

list of objects of class single_cancergram_pred. Each object of this class contains analyzed sequence, values of predictions for 5-mers and result of the prediction for the whole peptide/protein.

Details

CancerGram requires the external package, CancerGramModel, which contains models necessary to perform the prediction. The model can be installed using install_CancerGramModel.

Predictions for each protein are stored in objects of class single_cancergram_pred. It consists of three elements:

seq

Character vector of amino acid sequence of an analyzed peptide/protein

all_mers_pred

Matrix of predictions for each 5-mer (subsequence of 5 amino acids) of a sequence. Each row corresponds to one mer and columns to predicted classes (ACP, AMP or negative). Prediction value indicates probability that a 5-mer possesses anticancer activity (acp), antimicrobial activity (amp) or none of them (neg).

single_prot_pred

One row matrix of a single prediction value for a whole peptide/protein. Its value corresponds to the probability that a peptide/protein exhibits anticancer activity, antimicrobial activity or none of them.